Fix todo
