+++ client/internal/engine_stdnet.go
@@ -7,5 +7,5 @@ import (
 )
 
 func (e *Engine) newStdNet() (*stdnet.Net, error) {
-	return stdnet.NewNet(nil, e.config.IFaceBlackList)
+	return stdnet.NewNet(e.config.IFaceBlackList)
 }
